+++ src/Symfony/Component/DependencyInjection/Kernel/ServicesBundle.php
@@ -31,6 +31,7 @@ use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigura
 use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\DependencyInjection\ServiceLocator;
 use Symfony\Component\EventDispatcher\Attribute\AsEventListener;
+use Symfony\Component\EventDispatcher\DependencyInjection\CompileListenersPass;
 use Symfony\Component\EventDispatcher\DependencyInjection\RegisterListenersPass;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
@@ -52,6 +53,9 @@ class ServicesBundle extends AbstractBundle
     {
         if (class_exists(RegisterListenersPass::class)) {
             $container->addCompilerPass(new RegisterListenersPass(), PassConfig::TYPE_BEFORE_REMOVING);
+        }
+        if (class_exists(CompileListenersPass::class)) {
+            $container->addCompilerPass(new CompileListenersPass(), PassConfig::TYPE_AFTER_REMOVING);
         }
         // Must run before ResolveInstanceofConditionalsPass which consumes and removes the parameter
         $container->addCompilerPass(new AddBehaviorDescribingTagsPass([